Split MCP server into focused modules and apply formatting fixes
Split the monolithic server.rs (1548 lines) into four modules to comply
with the 1000-line limit: server.rs (run loop, 223 lines), types.rs
(request/response types and param structs, 605 lines), tools.rs (tool
specs and lookup, 309 lines), handlers.rs (method dispatch and handlers,
505 lines). All 38 tests preserved and passing. Also applies rustfmt
formatting fixes to generated CKAN model files.
